REFACTOR: move the checkRentPaid and calculateCondoFees functions to FinancialHandler.js for separation of concerns #189
